pyproject.toml
[project]
name = "chat-agents-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Huy Tran", email = "khanhhuy288@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "agent-framework[devui]>=1.0.0b251114",
    "aiohttp>=3.13.2",
    "azure-ai-projects==1.1.0b4",
    "azure-identity>=1.26.0b1",
    "fastapi>=0.115.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.21.2",
    "openai>=2.8.1",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "uvicorn[standard]>=0.32.0",
]

[project.scripts]
chat-agents-system-devui = "chat_agents_system.devui_app:main"
chat-agents-system-api = "chat_agents_system.api_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chat_agents_system"]
